move view property shortcuts to composition shortcuts
