+++ src/timeline/timelineShortcuts.ts
@@ -1,3 +1,4 @@
+import { compositionKeyboardShortcuts } from "~/composition/compositionShortcuts";
 import { reduceCompProperties } from "~/composition/compositionUtils";
 import { compSelectionFromState } from "~/composition/util/compSelectionUtils";
 import { AreaType } from "~/constants";
@@ -5,7 +6,7 @@ import { RequestActionParams } from "~/listener/requestAction";
 import { createOperation } from "~/state/operation";
 import { areaActionStateFromState, getActionState } from "~/state/stateUtils";
 import { timelineOperations } from "~/timeline/operations/timelineOperations";
-import { KeyboardShortcut, PropertyName, ShouldAddShortcutToStackFn } from "~/types";
+import { KeyboardShortcut, ShouldAddShortcutToStackFn } from "~/types";
 
 const getAreaActionState = (areaId: string, actionState = getActionState()) =>
 	areaActionStateFromState<AreaType.Timeline>(areaId, actionState);
@@ -31,7 +32,7 @@ const getSelectedTimelineIds = (areaId: string, actionState = getActionState()) 
 	return timelineIds;
 };
 
-const timelineShortcuts = {
+export const timelineShortcuts = {
 	removeSelectedKeyframes: (areaId: string, params: RequestActionParams) => {
 		const { compositionId } = getAreaActionState(areaId);
 		const timelineIds = getSelectedTimelineIds(areaId);
@@ -47,27 +48,6 @@ const timelineShortcuts = {
 
 		const op = createOperation();
 		timelineOperations.easyEaseSelectedKeyframes(op, timelineIds);
-
-		params.dispatch(op.actions);
-	},
-
-	viewTransformProperties: (propertyNames: PropertyName[]) => (
-		areaId: string,
-		params: RequestActionParams,
-	) => {
-		const { compositionId } = getAreaActionState(areaId);
-
-		const op = createOperation();
-		timelineOperations.viewTransformProperties(op, compositionId, propertyNames);
-
-		params.dispatch(op.actions);
-	},
-
-	viewAnimatedProperties: (areaId: string, params: RequestActionParams) => {
-		const { compositionId } = getAreaActionState(areaId);
-
-		const op = createOperation();
-		timelineOperations.viewAnimatedProperties(op, compositionId);
 
 		params.dispatch(op.actions);
 	},
@@ -94,6 +74,7 @@ const wereKeyframesRemoved: ShouldAddShortcutToStackFn = (areaId, prevState, nex
 };
 
 export const timelineKeyboardShortcuts: KeyboardShortcut[] = [
+	...compositionKeyboardShortcuts,
 	{
 		name: "Remove selected keyframes",
 		key: "Backspace",
@@ -110,37 +91,5 @@ export const timelineKeyboardShortcuts: KeyboardShortcut[] = [
 		name: "Easy ease selected keyframes",
 		key: "F9",
 		fn: timelineShortcuts.easeEaseSelectedKeyframes,
-	},
-	{
-		name: "View animated properties",
-		key: "U",
-		fn: timelineShortcuts.viewAnimatedProperties,
-	},
-	{
-		name: "View position property",
-		key: "P",
-		fn: timelineShortcuts.viewTransformProperties([
-			PropertyName.PositionX,
-			PropertyName.PositionY,
-		]),
-		optionalModifierKeys: ["Shift"],
-	},
-	{
-		name: "View rotation property",
-		key: "R",
-		fn: timelineShortcuts.viewTransformProperties([PropertyName.Rotation]),
-		optionalModifierKeys: ["Shift"],
-	},
-	{
-		name: "View scale property",
-		key: "S",
-		fn: timelineShortcuts.viewTransformProperties([PropertyName.ScaleX, PropertyName.ScaleY]),
-		optionalModifierKeys: ["Shift"],
-	},
-	{
-		name: "View anchor property",
-		key: "A",
-		fn: timelineShortcuts.viewTransformProperties([PropertyName.AnchorX, PropertyName.AnchorY]),
-		optionalModifierKeys: ["Shift"],
 	},
 ];
